fast-sim/pyproject.toml
[project]
name = "fast-sim"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35",
    "bedrock-agentcore>=0.1.0",
    "pyarrow>=17.0",
    "aws-cdk-lib>=2.170",
    "constructs>=10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
    "pyright>=1.1",
    "pytest>=8.0",
    "moto[all]>=5.0",
]

[project.scripts]
fast-sim = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"
exclude = ["infra/"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
